started using flags; also ERROR: rotaryencoder causes mutex in ISR

+++ My-Servo-Lib/myServo.h
@@ -27,5 +27,5 @@ private:
 
     float position;
 
-    uint8_t positionDegrees;
+    float positionDegrees;
 };
